+++ eventmesh-test/src/main/java/org/apache/eventmesh/http/demo/SyncRequestInstance.java
@@ -1,34 +1,26 @@
-package org.apache.eventmesh.eventmesh.http.demo;
+package org.apache.eventmesh.http.demo;
 import com.webank.eventmesh.client.http.conf.LiteClientConfig;
 import com.webank.eventmesh.client.http.producer.LiteProducer;
-import com.webank.eventmesh.client.http.producer.RRCallback;
 import com.webank.eventmesh.common.IPUtil;
 import com.webank.eventmesh.common.LiteMessage;
 import com.webank.eventmesh.common.ThreadUtil;
-import org.apache.eventmesh.eventmesh.util.Utils;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Properties;
+public class SyncRequestInstance {
 
-public class AsyncSyncRequestInstance {
-
-    public static Logger logger = LoggerFactory.getLogger(AsyncSyncRequestInstance.class);
+    public static Logger logger = LoggerFactory.getLogger(SyncRequestInstance.class);
 
     public static void main(String[] args) throws Exception {
 
-        Properties properties = Utils.readPropertiesFile("application.properties");
-        final String eventMeshIp = properties.getProperty("eventmesh.ip");
-        final String eventMeshHttpPort = properties.getProperty("eventmesh.http.port");
-
         LiteProducer liteProducer = null;
         try {
-//            String eventMeshIPPort = args[0];
-            String eventMeshIPPort = eventMeshIp + ":" + eventMeshHttpPort;
-//            final String topic = args[1];
-            final String topic = "FT0-e-80010000-01-1";
+            String eventMeshIPPort = args[0];
+
+            final String topic = args[1];
+
             if (StringUtils.isBlank(eventMeshIPPort)) {
                 // if has multi value, can config as: 127.0.0.1:10105;127.0.0.2:10105
                 eventMeshIPPort = "127.0.0.1:10105";
@@ -44,29 +36,21 @@ public class AsyncSyncRequestInstance {
                     .setPid(String.valueOf(ThreadUtil.getPID()));
 
             liteProducer = new LiteProducer(eventMeshClientConfig);
+            liteProducer.start();
 
-            final long startTime = System.currentTimeMillis();
-            final LiteMessage liteMessage = new LiteMessage();
+            long startTime = System.currentTimeMillis();
+            LiteMessage liteMessage = new LiteMessage();
             liteMessage.setBizSeqNo(RandomStringUtils.randomNumeric(30))
-                    .setContent("testAsyncMessage")
+                    .setContent("contentStr with special protocal")
                     .setTopic(topic)
                     .setUniqueId(RandomStringUtils.randomNumeric(30));
 
-            liteProducer.request(liteMessage, new RRCallback() {
-                @Override
-                public void onSuccess(LiteMessage o) {
-                    logger.debug("sendmsg : {}, return : {}, cost:{}ms", liteMessage.getContent(), o.getContent(), System.currentTimeMillis() - startTime);
-                }
-
-                @Override
-                public void onException(Throwable e) {
-                    logger.debug("sendmsg failed", e);
-                }
-            }, 3000);
-
-            Thread.sleep(2000);
+            LiteMessage rsp = liteProducer.request(liteMessage, 10000);
+            if (logger.isDebugEnabled()) {
+                logger.debug("sendmsg : {}, return : {}, cost:{}ms", liteMessage.getContent(), rsp.getContent(), System.currentTimeMillis() - startTime);
+            }
         } catch (Exception e) {
-            logger.warn("async send msg failed", e);
+            logger.warn("send msg failed", e);
         }
 
         try{
