+++ eventmesh-test/src/main/java/org/apache/eventmesh/http/demo/sub/SpringBootDemoApplication.java
@@ -1,4 +1,4 @@
-package org.apache.eventmesh.eventmesh.http.demo.sub;
+package org.apache.eventmesh.http.demo.sub;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
