+++ eventmesh-test/src/main/java/org/apache/eventmesh/http/demo/sub/service/SubService.java
@@ -1,11 +1,11 @@
-package org.apache.eventmesh.eventmesh.http.demo.sub.service;
+package org.apache.eventmesh.http.demo.sub.service;
 
 import com.webank.eventmesh.client.http.conf.LiteClientConfig;
 import com.webank.eventmesh.client.http.consumer.LiteConsumer;
 import com.webank.eventmesh.common.IPUtil;
 import com.webank.eventmesh.common.EventMeshException;
 import com.webank.eventmesh.common.ThreadUtil;
-import org.apache.eventmesh.eventmesh.util.Utils;
+import org.apache.eventmesh.util.Utils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
