+++ eventmesh-test/src/main/java/org/apache/eventmesh/tcp/common/EventMeshTestCaseTopicSet.java
@@ -1,4 +1,4 @@
-package org.apache.eventmesh.eventmesh.tcp.common;
+package org.apache.eventmesh.tcp.common;
 
 /**
  * Testcase set
