+++ eventmesh-test/src/main/java/org/apache/eventmesh/tcp/demo/AsyncPublish.java
@@ -1,12 +1,12 @@
-package org.apache.eventmesh.eventmesh.tcp.demo;
+package org.apache.eventmesh.tcp.demo;
 
 import com.webank.eventmesh.client.tcp.EventMeshClient;
 import com.webank.eventmesh.client.tcp.common.EventMeshCommon;
 import com.webank.eventmesh.client.tcp.impl.DefaultEventMeshClient;
 import com.webank.eventmesh.common.protocol.tcp.Package;
 import com.webank.eventmesh.common.protocol.tcp.UserAgent;
-import org.apache.eventmesh.eventmesh.tcp.common.EventMeshTestUtils;
-import org.apache.eventmesh.eventmesh.util.Utils;
+import org.apache.eventmesh.tcp.common.EventMeshTestUtils;
+import org.apache.eventmesh.util.Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
